+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Fixed
+-  Prometheus target metrics scraper will no longer panic if a metric has an
+  invalid value (instead it will be logged)
 
 ## [0.23.3]
 ### Changed